+++ dandy/bot/bot.py
@@ -3,5 +3,6 @@ from abc import ABC
 from dandy.processor.processor import BaseProcessor
 
 
-class Bot(BaseProcessor, ABC):
+class BaseBot(BaseProcessor, ABC):
     ...
+