+++ dandy/workflow/workflow.py
@@ -3,5 +3,5 @@ from abc import ABC
 from dandy.processor.processor import BaseProcessor
 
 
-class Workflow(BaseProcessor, ABC):
+class BaseWorkflow(BaseProcessor, ABC):
     ...