+++ tests/bot/test_bot.py
@@ -5,5 +5,5 @@ from dandy.processor.processor import BaseProcessor
 
 class TestBot(TestCase):
     def test_bot_import(self):
-        from dandy.bot import Bot
-        self.assertTrue(type(Bot) is type(BaseProcessor))
+        from dandy.bot import BaseBot
+        self.assertTrue(type(BaseBot) is type(BaseProcessor))