+++ tests/workflow/test_workflow.py
@@ -5,5 +5,5 @@ from dandy.processor.processor import BaseProcessor
 
 class TestWorkflow(TestCase):
     def test_workflow_import(self):
-        from dandy.workflow import Workflow
-        self.assertTrue(type(Workflow) is type(BaseProcessor))
+        from dandy.workflow import BaseWorkflow
+        self.assertTrue(type(BaseWorkflow) is type(BaseProcessor))